Update AppServiceProvider.php
This line made the local server try to run under the path /ghap/ Because of this it could not find the css and other files that were in the root path.  I think must be a legacy from before we consolidated TLCMap to simply be GHAP, so am simply removing the /ghap subdirectory from this path. This is also just a very small change to check commit and push works ok after setting up local dev environment.