Upgrade to Bootstrap 5.3.3 for better responsive green navbar.
To make a better responsive navbar for small screen mobile devices it seemed best to upgrade to bootstrap 5.3.3. The improvement to the navbar was mainly to put the burger in the right place in the drop down, and remove the big menu on small screen. There may be some small tweaks to appearance still . This is part of the 'Fieldwork' changes for the Nukunu project. It's better to make the screens more responsive, and simpler for sm, than to create a separate interface.